requirements.txt
fastapi==0.120.1
uvicorn[standard]==0.23.2
sqlalchemy==2.0.20
psycopg2-binary==2.9.8
pydantic[email]==1.10.12
email-validator==1.3.1
python-dotenv==1.2.1